% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewars.R
\name{ewars_fetch_data}
\alias{ewars_fetch_data}
\title{Title}
\usage{
ewars_fetch_data(
  ewars_config,
  start_date = NULL,
  end_date = lubridate::today(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{}
}
\description{
Title
}
