% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewars.R
\name{ewars_login_dhis2}
\alias{ewars_login_dhis2}
\title{Title}
\usage{
ewars_login_dhis2(
  config_path = NULL,
  config_path_level = "dhis",
  username = NULL,
  password = NULL,
  base_url = NULL,
  d2_session_name = "d2_default_session",
  d2_session_envir = parent.frame()
)
}
\arguments{
\item{d2_session_envir}{}
}
\description{
Title
}
